05.11 UserReservationService updateCurrentStatus
