+++ src/main/java/com/project/ihealme/HptReception/controller/HptReceptionController.java
@@ -41,13 +41,13 @@ public class HptReceptionController {
         return "success";
     }
 
-    @GetMapping("/HptReception/HptReceptionList/updateCurrentStatusToAccept")       // <a> 태그는 get 방식으로 요청한다.
-    public String updateCurrentStatusToAccept(@RequestParam("resNo") int resNo) {
-        hptReceptionService.updateCurrentStatus(resNo, "진료 전", LocalDateTime.now());
-        hptReceptionService.increaseRtCount(); // 대기자 수 +1
-
-        return "redirect:/HptReception/HptReceptionList";
-    }
+//    @GetMapping("/HptReception/HptReceptionList/updateCurrentStatusToAccept")       // <a> 태그는 get 방식으로 요청한다.
+//    public String updateCurrentStatusToAccept(@RequestParam("resNo") int resNo) {
+//        hptReceptionService.updateCurrentStatus(resNo, "진료 전", LocalDateTime.now());
+//        hptReceptionService.increaseRtCount(); // 대기자 수 +1
+//
+//        return "redirect:/HptReception/HptReceptionList";
+//    }
 
 //    @GetMapping("/HptReception/HptReceptionList/updateCurrentStatusToReject")
 //    public String updateCurrentStatusToReject(@RequestParam("resNo") int resNo) {
