+++ src/main/java/com/project/ihealme/userReservation/domain/UserReservation.java
@@ -33,11 +33,11 @@ public class UserReservation extends BaseEntity {
     @ColumnDefault("'접수 대기'")
     private String currentStatus;
 
-    public UserReservation toEntity(UserReservation userRes) {
-        UserReservation userReservation = userRes.builder()
-                .resNo(resNo)
-                .currentStatus("접수취소")
-                .build();
-        return userReservation;
-    }
+//    public UserReservation toEntity(UserReservation userRes) {
+//        UserReservation userReservation = userRes.builder()
+//                .resNo(resNo)
+//                .currentStatus("접수취소")
+//                .build();
+//        return userReservation;
+//    }
 }
