+++ src/main/resources/static/css/reservationStyle.css
@@ -1,5 +1,19 @@
 @charset "UTF-8";
 
+nav {
+  width: 100%;
+  height: 57px;
+  display: flex;
+  align-items: center;
+}
+
+header {
+  width: 100%;
+  height: 73px;
+  display: flex;
+  align-items: center;
+}
+
 body {
 	font-family: Arial, sans-serif;
 	margin: 0;
